refactor: no-logic readers moved to new readers sub-package
